Use GCS exports for audit log analysis if available (#230)
When analyzing a project, check if there are applicable GCS audit log 
exports available. If so, prefer the export over querying the API as
it is likely going to be faster.

An export is considered applicable if
* the associated sink is active and has been around for long enough to
  cover the entire analysis period
* The associated bucket is accessible 
* The bucket contains at least one admin audit log export object

+++ sources/Google.Solutions.Common/Util/LinqExtensions.cs
@@ -43,7 +43,7 @@ namespace Google.Solutions.Common.Util
             return !lookup.Except(sequence).Any();
         }
 
-        public static IEnumerable<IEnumerable<T>> Chunk<T>(this IEnumerable<T> source, int chunkSize)
+        public static IEnumerable<IEnumerable<T>> Chunk<T>(this IEnumerable<T> source, ushort chunkSize)
         {
             var chunk = new List<T>(chunkSize);
             foreach (var x in source)
